# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD API Test",
        "description": "تست API در لاراول",
        "contact": {
            "email": "vreyazat@gmail.com"
        },
        "license": {
            "name": "Vahid Reyazat",
            "url": ""
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "گرفتن کل داده های مشتری ها",
                "description": "گرفتن کل داده های مشتری ها",
                "operationId": "da9547eb734ff6976718d5328940c13a",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json;",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "گرفتن کل داده های مشتری ها",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "گرفتن کل داده های مشتری ها": {
                                        "summary": "گرفتن کل داده های مشتری ها",
                                        "value": {
                                            "status": "success",
                                            "code": 200,
                                            "message": "Successfully logged out",
                                            "errors": [],
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "گرفتن کل داده های مشتری ها",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "گرفتن کل داده های مشتری ها": {
                                        "summary": "گرفتن کل داده های مشتری ها",
                                        "value": {
                                            "status": "error",
                                            "code": 401,
                                            "message": "Unauthenticated.",
                                            "errors": [],
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "حذف مشتری ",
                "description": "حذف مشتری",
                "operationId": "19467afa9b9eacaa9de17f2be03a098f",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json;",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "حذف مشتری",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "حذف مشتری": {
                                        "summary": "حذف مشتری",
                                        "value": {
                                            "status": "success",
                                            "code": 200,
                                            "message": "The item was deleted!",
                                            "errors": [],
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}